refactor(data): reorganize UnionBan data processing logic
