+++ src/main/java/fun/blue_millennium/module/impl/UnionBan.java
@@ -7,7 +7,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 
-import static fun.blue_millennium.data.UnionBan.LocalBanListImport.importBanList;
+import static fun.blue_millennium.data.UnionBan.LocalProcess.LocalBanListImport.importBanList;
 import static fun.blue_millennium.data.UnionBan.LocalProcess.OnlineDataMerge.mergeAndReportData;
 
 public class UnionBan extends Module implements Listener {
